+++ lib/src/models/matches/index.dart
@@ -40,3 +40,4 @@ export 'segments.dart';
 export 'segment_metadata.dart';
 export 'segment_attributes.dart';
 export 'kd_histogram.dart';
+export 'stats.dart';