+++ lib/src/models/profile/metadata.dart
@@ -43,8 +43,10 @@ class OtherTitles {
 
   OtherTitles({this.warzone});
 
-  factory OtherTitles.fromMap(Map<String, dynamic> json) =>
-      OtherTitles(warzone: json.containsKey('warzone') ? Warzone.fromMap(json['warzone']) : null );
+  factory OtherTitles.fromMap(Map<String, dynamic> json) => OtherTitles(
+      warzone: json.containsKey('warzone')
+          ? Warzone.fromMap(json['warzone'])
+          : null);
 
   Map<String, dynamic> toMap() => {'warzone': warzone?.toMap()};
 }